requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.0
kiwisolver==1.4.4
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
PyOpenGL==3.1.7
pyparsing==3.0.9
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
